all xetra error checked

+++ part1/ft_strncmp.c
@@ -1,6 +1,9 @@
+#include "libft.h"
+
+
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+    size_t i;
     int difference;
 
     i = 0;
